requirements.txt
streamlit
pandas
pillow
requests
fpdf